+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Add `salmon.util/pages-seq` for handling paginated responses.
+- Add `salmon.cleanup/deregister-all-amis!` for cleaning up AMIs in test accounts.
 
 ## v0.19.0 (2024-09-16)